Refactor base Graphene object types
