+++ netbox/netbox/graphql/types.py
@@ -1,7 +1,7 @@
-import graphene
 from django.contrib.contenttypes.models import ContentType
-from graphene.types.generic import GenericScalar
 from graphene_django import DjangoObjectType
+
+from extras.graphql.mixins import CustomFieldsMixin, TagsMixin
 
 __all__ = (
     'BaseObjectType',
@@ -27,30 +27,20 @@ class BaseObjectType(DjangoObjectType):
         return queryset.restrict(info.context.user, 'view')
 
 
-class ObjectType(BaseObjectType):
+class ObjectType(CustomFieldsMixin, BaseObjectType):
     """
-    Extends BaseObjectType with support for custom field data.
+    Extends BaseObjectType with support for custom fields.
     """
-    custom_fields = GenericScalar()
-
     class Meta:
         abstract = True
 
-    def resolve_custom_fields(self, info):
-        return self.custom_field_data
 
-
-class TaggedObjectType(ObjectType):
+class TaggedObjectType(CustomFieldsMixin, TagsMixin, BaseObjectType):
     """
-    Extends ObjectType with support for Tags
+    Extends BaseObjectType with support for custom fields and tags
     """
-    tags = graphene.List(graphene.String)
-
     class Meta:
         abstract = True
-
-    def resolve_tags(self, info):
-        return self.tags.all()
 
 
 #
